Gary's 2017-12-03 tarball, untarred on top of his previous (11/28) tarball.
I've reworked the branches a little to try to make it easier to integrate
new (non-git) tarballs into the pending git-ified work.  The idea is ithat
the "gary" branch will just hold successive tarballs, which I will then
merge into the git-ified code in other branches.  The hope is that this
will avoid having the tarbals simply undoing recent work, letting merge be
smarter.  This, I've dropped that "gary" branch back to the point of his
previous tarball, instead of doing any merges in this branch.