Assimilated tarball from email of 2017-12-11.
changes in directory structure include:

-- tests are moved to test/

-- old stuff that Gary had moved to test/old is not in this commit

-- gary added a tarball of the testdir data-subdir, used as a standard set
of inputs for test scripts.  Such a tarball might make sense to use, going
forward, because it can contain standardized UID/GID, instead of just
matching whoever checked out the repo.  Still work to be done on that,
though, so I also kept the old directory contents as part of this commit.