Gary's tarball of 2017-12-11
Gary's work was done on top of a tarball made from Jeff's git work, so it
contains the C-Thread-Pool submodule as just files, plus a .gitmodules
file.  That's not what a real submodule looks like.  I'll ignore the
submodule trappings in the tarball, and try explicitly adding C-Thread-Pool
as a new submodule here, then committing the new .modules and C-Thread-Pool
dir.  Might be a mistake, as the purpose of this branch is just to receive
tarballs for merging into other branches.  Hoping that merging submodules
is smart.

There is also a lot of file movement here, shunting stuff off into test/
and scripts/.  A bunch of cruft was moved into misc/, but I think that that
stuff probably doesn't want to be in git any longer, so I've left it out.