Preliminary build of GUFI fuse-support on Linux.
This is the only place I've been able to test, so far.
I hope I haven't broken things for the OSX build.

run 'make help' for options.  The default doesn't attempt to
build fuse.  Use 'make fuse' or 'make everything' to build fuse.

Updated test/runbfqforfuse to work for any user, though it's not yet
running to completion, for me.  Fixed bfresultfuse to report an error
instead of exiting, at the problem spot.