Fixed duplicate inode printing and .lastrun time
The previous version of the GPFS scan tool had an error in proc_inodes where threads would print duplicate inodes.
This has been fixed. Additionally, the .lastrun file created for incremental scans had been created with a time
from when the scan finished. Now it has the time when the scan starts so we don't miss changes that occur during
the scan.