Added individual OutputBuffer mutexes
A global mutex for a OutputBuffers creates contention between threads writing to their own OutputBuffer, so OutputBuffer was given a mutex. However, if no output file prefix is provided, all of the OutputBuffers will write to the same file pointer, which is a race condition. A global mutex is created if no output file prefix is provided to prevent this.

gufi_dir2trace still has the race condition
    writes are done directly to the output file, not OutputBuffers - need to change
    simply using mutexes requires wrapping the entire loop of processdir, which causes too much contention

OutputBuffer and OutputBuffers api rework
    OutputBuffer_flush no longer locks since it does not care where it does not know who else is accessing the file
    Added OutputBuffer_write to do memcpys like in gufi_query
    OutputBuffers_* after init do not take in the size any more - it is stored in the struct

Added OutputBuffer and OutputBuffers tests