Allow for top-level input paths to gufi_query to be symlinks
This is intended to facilitate deployment of multiple sets of indexes
by allowing for the top-level search directories that contains
multiple indexes to be symlinked, allowing for atomic pivots to new
sets of indexes:

    # create indexes at timestep 0
    gufi_dir2index index0 /timestep0
    gufi_dir2index index1 /timestep0
    ...

    # expose the latest indexes to users
    ln -s /timestep0 /search

    # create indexes at timestep 1
    gufi_dir2index index0 /timestep1
    gufi_dir2index index1 /timestep1
    ...

    # pivot to timestep 1 using the same user path
    #
    # already started queries continue to walk /timestep0
    # new queries walk /timestep1
    ln -snf /timestep1 /search

Side effect: Paths passed into gufi_query that are below the root can
also be symlinks. However, if running a query from a path above the
symlink, gufi_query will ignore the symlink when it is reached.

Custom SQLite functions path and rpath now show strings with the user
input path, not the real path (if the user-provided starting path was
a symlink).
    fpath prints the real path

test fixes
added simple deployment example/test