steal work that has already been claimed
QueuePerThreadPool normally claims an entire work queue at once
instead of popping them off one at a time. This reduces the contention
on the work queues. However, this scheme makes it possible to cause
starvation when a work item takes a long time to complete and has more
work items queued up after it, all of which have already been claimed
and are not in the waiting or deferred queues.

Added steal_active function to prevent starvation by stealing work
that has already been claimed, reproducing the effect of threads
popping off individual work items. steal_active is only called after
the waiting and deferred queues have been checked and found to be
empty, and so should not be called frequently.