handle errors in zlib uncompress()
Currently errors in zlib's uncompress() function are ignored, which
in the worst case could mean continuing to run the program with memory
corruption or unexpected results.

There are two error conditions that can be encountered. This patch
makes the program abort if it encounters either.

- Running out of memory: while this is a valid condition and could be
  handled, the program is unlikely to be able to make any useful
  progress if it cannot allocate memory, and aborting means we don't
  have to add error handling code for this case.

- Invalid or corrupt data: because this program is the source of all
  data that it uncompresses, it should never encounter an invalid or
  corrupt zlib stream. If it does, that indicates a memory bug, and
  panicking is the safest course of action.

Because decompress_struct() now aborts on all error conditions, it is
changed to return void. This simplifies callers of this code as they
do not have to worry about handling error cases.