remove sll_move() in favor of sll_move_append()
sll_move() and sll_move_first() are easy to misuse because they drop
the contents of dst. This means in order to be used safely, the caller
must either ensure that dst is empty, or retain a reference to the
contents of dst itself and later process and free the contents in order
to avoid leaking them.

In practice, the only callers of sll_move[_first]() ensured that the
dst was empty, so there was no situation where the contents could be
leaked.

However, when dst is empty, sll_move() is equivalent to
sll_move_append() and the latter is safer as it is not possible to leak
list nodes with it.

Therefore, replace all calls to sll_move() with sll_move_append() since
the behavior is equivalent and we can then remove sll_move(), thus
simplifying the SLL API and making it harder to misuse.