make sll methods with unused return values return void
The singly linked list methods sll_push(), sll_move_append(), and
sll_move_append_first() return a reference to the list that they operate
on. However, this return value is not needed as it does not convey any
information to the caller.

These functions are called only for their side effects, and in practice
all callers (outside of the test suite) ignored the return value.

Making these functions return void simplifies the API and makes it clear
that they are only used for side effects and no information is conveyed
by the return value.