remove embedded path bufs to make `struct BottomUp` smaller
This makes `struct BottomUp` smaller by replacing two fixed-size buffers
of 4 KiB for pathnames with dynamically allocated buffers that are only
as large as they need to be to store the given path.

This reduces the size of `struct BottomUp` from 8504 bytes to 328 bytes.

I found that parallel_rmr used 1131492 KiB of virtual memory to delete a
tree of 55987 dirs and 111972 files before this change, and 63916 KiB
after this change.