only format non-directory path names when needed
Previously, bottomup would always format full pathnames for
non-directory nodes as they are encountered, even if those nodes are
never processed. This incurs an unnecessary allocation and format
operation.

It also meant that additional code had to be added to free
those pathname buffers since they weren't free'd by the normal code path
that processes nodes (since they aren't being processed).

With this change, bottomup checks the file type using the dirent
(falling back to fstatat(2) if needed), so that we can skip formatting
the child's name when it is a non-dir and we aren't processing non-dirs.