run_vt
This is intended for power users to be able to run arbitrary
commands and generate virtual tables with aribtrary result
columns. In order to achieve this, users MUST run "CREATE VIRTUAL
TABLE" to create the virtual table before querying it. The virtual
table should preferrably be created in the temp namespace:

    CREATE VIRTUAL TABLE temp.output
    USING run_vt(cmd='find -name "*.c" -printf "%f %s\n"', cols='name, size');

All arguments should be key=value pairs. The allowed keys are:
    cmd     - the command string
    cols    - a comma delimited list of column names (no types)
    d or D  - column delimiter (default: space)

Notes:
    - Arguments may appear in any order.
    - If there are duplicate arguments, the right-most duplicate will be used

Then, SELECT from the virtual table

    SELECT * FROM output;