set CMAKE_ENABLE_EXPORTS when building GUFI binaries
This corresponds to the `-export-dynamic` linker option.

This is needed for dynamically loading plugins that use sqlite3
functions.

GUFI binaries are built with the sqlite3 object file statically linked.
When a plugin is loaded that calls sqlite3 functons, it must refer to
the definitions of the sqlite3 symbols that live in the GUFI binary that
the plugin is loaded into.

The `-export-dynamic` linker option makes the GUFI binaries export these
symbols so that the plugin is able to resolve them correctly.