gufi_dir2index: remove unnecessary manipulation of entry->name
`entry->name` has the full path, but was being overwritten with just the
relative path. This isn't actually necessary for any GUFI functionality.

Previously it was harmless, but it will be problematic for handing the
file to a plugin to process, since the plugin will need the full path.

So in preparation for hooking in the plugin, get rid of the modification
to entry->name.